refactor(#430): extract shared lightbox into utils/lightbox.js
~100 LOC of lightbox state and logic was duplicated verbatim between
travel.js and travel-post.js. Any bug fix had to be applied twice.
Extracted to resources/js/utils/lightbox.js (exports openLightbox,
initLightbox) and replaced both copies with a single import. Added
the missing null guard from travel-post.js to the shared version.
